// types/text.ts
export type TextType =
  | "ORIGINAL_FILOSOFIA"
  | "FILOSOFIA"
  | "ORIGINAL_SOSTENIMIENTO_M"
  | "SOSTENIMIENTO_M"
  | "ORIGINAL_SOSTENIMIENTO_F"
  | "SOSTENIMIENTO_F"
  | "ORIGINAL_COMPROMISO_M"
  | "COMPROMISO_M"
  | "ORIGINAL_COMPROMISO_F"
  | "COMPROMISO_F"
  | "ORIGINAL_COMPROMISO_INTERNACION_M"
  | "COMPROMISO_INTERNACION_M"
  | "ORIGINAL_COMPROMISO_INTERNACION_F"
  | "COMPROMISO_INTERNACION_F";

export type Text = {
  id: number;
  name: TextType;
  textHtml: string;
  created_at: string;
};